Move db_mutex into utils/
