+++ metaspace/engine/sm/engine/annotation_lithops/moldb_pipeline.py
@@ -26,7 +26,7 @@ from sm.engine.annotation_lithops.io import (
     deserialize,
 )
 from sm.engine.annotation_lithops.utils import jsonhash
-from sm.engine.db_mutex import DBMutex
+from sm.engine.utils.db_mutex import DBMutex
 from sm.engine.ds_config import DSConfig
 from sm.engine.isocalc_wrapper import IsocalcWrapper
